Split NLG.hs in two modules
